# api/routes/hotels-canonical.js
/**
 * Canonical Hotel API Endpoints (STEP 2)
 *
 * Four main endpoints for hotel search and booking:
 * 1. GET /api/hotels/autocomplete - city/destination autocomplete
 * 2. POST /api/hotels/search - hotel search with dates and guests
 * 3. GET /api/hotels/:propertyId - hotel details by property ID
 * 4. POST /api/hotels/:propertyId/rates - get available room rates
 *
 * Design: TBO-first with supplier-agnostic schema
 * - All queries filter supplier_code = 'TBO' for STEP 2
 * - Rate caching with 15-minute TTL (configurable via ROOM_OFFER_TTL_MINUTES)
 * - Graceful fallback: return hotel content even if TBO fails
 * - pricing_available flag when rates unavailable
 *
 * Database tables:
 * - hotel_unified: canonical hotel master
 * - room_offer_unified: cached room rates with TTL
 * - hotel_images: gallery images (if available)
 */

const express = require("express");
const { v4: uuidv4 } = require("uuid");
const db = require("../database/connection");
const supplierAdapterManager = require("../services/adapters/supplierAdapterManager");
const router = express.Router();

// Configuration
const ROOM_OFFER_TTL_MINUTES = parseInt(
  process.env.ROOM_OFFER_TTL_MINUTES || "15",
);
const USE_SUPPLIER_FILTER = "TBO"; // STEP 2: TBO only

/**
 * Utility: Get TBO adapter instance
 */
function getTboAdapter() {
  const adapter = supplierAdapterManager.getAdapter("TBO");
  if (!adapter) {
    throw new Error("TBO adapter not initialized");
  }
  return adapter;
}

/**
 * Utility: Build TTL expiry timestamp
 */
function getExpiryTimestamp() {
  return new Date(Date.now() + ROOM_OFFER_TTL_MINUTES * 60 * 1000);
}

/**
 * Utility: Extract numeric city ID from city code
 */
function extractCityId(cityCode) {
  // TBO uses city codes like "DXB", "PAR", etc.
  // Some APIs expect numeric IDs
  // For now, pass as-is; enrichment happens in adapter
  return cityCode;
}

/**
 * ============================================================================
 * ENDPOINT 1: GET /api/hotels/autocomplete
 * Destination/city autocomplete for search form
 * ============================================================================
 */
router.get("/autocomplete", async (req, res) => {
  try {
    const { q = "", limit = 15, country = null } = req.query;

    if (!q || q.trim().length === 0) {
      return res.status(400).json({
        success: false,
        error: 'Query parameter "q" is required',
        suggestions: [],
      });
    }

    const adapter = getTboAdapter();

    // Call TBO adapter's searchCities method
    // Adapter handles ranking (starts-with > contains) and caching
    const cities = await adapter.searchCities(
      q.trim(),
      Math.min(parseInt(limit) || 15, 100),
      country,
    );

    // Transform to canonical response format
    const suggestions = cities.map((city) => ({
      city_code: city.code || city.cityCode || "",
      city_name: city.name || city.cityName || "",
      country_code: city.countryCode || city.country_code || "",
      country_name: city.countryName || city.country_name || "",
      type: city.type || "CITY",
      lat: city.latitude || city.lat || null,
      lng: city.longitude || city.lng || null,
    }));

    res.json({
      success: true,
      suggestions: suggestions.slice(0, parseInt(limit) || 15),
    });
  } catch (error) {
    console.error("❌ Hotels autocomplete error:", error.message);
    res.status(500).json({
      success: false,
      error: "Unable to fetch suggestions",
      suggestions: [],
    });
  }
});

/**
 * ============================================================================
 * ENDPOINT 2: POST /api/hotels/search
 * Hotel search with dates, guests, and optional filters
 * ============================================================================
 */
router.post("/search", async (req, res) => {
  try {
    const {
      city_code, // required: destination city code (DXB, PAR, etc.)
      country_code, // optional: country code for city lookup
      check_in, // required: YYYY-MM-DD format
      check_out, // required: YYYY-MM-DD format
      adults = 2,
      children = 0,
      rooms = 1,
      guest_nationality = "IN",
      preferred_currency = "INR",
      limit = 50,
      offset = 0,
    } = req.body;

    // Validation
    if (!city_code || !check_in || !check_out) {
      return res.status(400).json({
        success: false,
        error: "Missing required fields: city_code, check_in, check_out",
      });
    }

    // Verify dates
    const checkInDate = new Date(check_in);
    const checkOutDate = new Date(check_out);
    if (isNaN(checkInDate.getTime()) || isNaN(checkOutDate.getTime())) {
      return res.status(400).json({
        success: false,
        error: "Invalid date format. Use YYYY-MM-DD",
      });
    }
    if (checkInDate >= checkOutDate) {
      return res.status(400).json({
        success: false,
        error: "Check-in must be before check-out",
      });
    }

    const nights = Math.ceil(
      (checkOutDate - checkInDate) / (1000 * 60 * 60 * 24),
    );

    // Search via TBO adapter
    const adapter = getTboAdapter();
    let hotels = [];
    let tboError = null;

    try {
      // Call TBO search
      const tboSearchParams = {
        cityId: extractCityId(city_code),
        countryCode: country_code,
        checkInDate: check_in,
        checkOutDate: check_out,
        adults: parseInt(adults),
        children: parseInt(children),
        rooms: parseInt(rooms),
        guestNationality: guest_nationality,
        preferredCurrency: preferred_currency,
      };

      console.log("🔍 TBO hotel search:", tboSearchParams);

      const tboResults = await adapter.searchHotels(tboSearchParams);
      hotels = Array.isArray(tboResults) ? tboResults : [];

      console.log(`✅ TBO returned ${hotels.length} hotels`);
    } catch (error) {
      tboError = error;
      console.error("⚠️ TBO search failed:", error.message);
      // Graceful fallback: return hotel content from DB without pricing
    }

    // Fetch hotel master data for all results
    let hotelRecords = [];
    try {
      if (hotels.length > 0) {
        // If TBO returned results, use those hotel IDs
        const hotelCodes = hotels
          .map((h) => h.hotelCode || h.HotelCode)
          .filter(Boolean)
          .slice(0, parseInt(limit));

        const result = await db.query(
          `SELECT property_id, supplier_code, supplier_hotel_id, hotel_name, address, city, country,
                  star_rating, review_score, review_count, thumbnail_url, amenities_json, lat, lng
           FROM hotel_unified
           WHERE supplier_code = $1 AND supplier_hotel_id = ANY($2)
           LIMIT $3 OFFSET $4`,
          [USE_SUPPLIER_FILTER, hotelCodes, parseInt(limit), parseInt(offset)],
        );

        hotelRecords = result.rows;
      }
    } catch (error) {
      console.error("⚠️ Failed to fetch hotel records:", error.message);
    }

    // If no TBO results and no DB records, return empty
    if (hotels.length === 0 && hotelRecords.length === 0) {
      return res.json({
        success: true,
        hotels: [],
        total: 0,
        pricing_available: false,
        message: tboError
          ? "Pricing temporarily unavailable from supplier"
          : "No hotels found",
      });
    }

    // Enrich with room rates
    const enrichedHotels = await Promise.all(
      hotelRecords.map(async (hotelRec) => {
        let rates = [];
        let pricingAvailable = true;

        if (!tboError) {
          try {
            // Fetch cached rates from room_offer_unified
            const rateResult = await db.query(
              `SELECT * FROM room_offer_unified
               WHERE property_id = $1 
               AND supplier_code = $2
               AND search_checkin = $3::date
               AND search_checkout = $4::date
               AND expires_at > NOW()
               ORDER BY price_total ASC
               LIMIT 10`,
              [hotelRec.property_id, USE_SUPPLIER_FILTER, check_in, check_out],
            );

            rates = rateResult.rows || [];
          } catch (error) {
            console.error("⚠️ Failed to fetch room rates:", error.message);
            pricingAvailable = false;
          }
        } else {
          pricingAvailable = false;
        }

        // Fetch images from hotel_images table
        let images = [];
        try {
          const imgResult = await db.query(
            `SELECT image_url FROM hotel_images
             WHERE property_id = $1
             ORDER BY "order" ASC
             LIMIT 5`,
            [hotelRec.property_id],
          );
          images = imgResult.rows ? imgResult.rows.map((r) => r.image_url) : [];
        } catch (error) {
          // Fallback to thumbnail if available
          if (hotelRec.thumbnail_url) {
            images = [hotelRec.thumbnail_url];
          }
        }

        // Format response
        const cheapestRate = rates[0];
        return {
          property_id: hotelRec.property_id,
          supplier_code: hotelRec.supplier_code,
          supplier_hotel_id: hotelRec.supplier_hotel_id,
          hotel_name: hotelRec.hotel_name,
          address: hotelRec.address,
          city: hotelRec.city,
          country: hotelRec.country,
          star_rating: parseFloat(hotelRec.star_rating) || null,
          review_score: parseFloat(hotelRec.review_score) || null,
          review_count: hotelRec.review_count || 0,
          images: images,
          amenities: hotelRec.amenities_json || [],
          location: {
            lat: parseFloat(hotelRec.lat) || null,
            lng: parseFloat(hotelRec.lng) || null,
          },
          pricing:
            pricingAvailable && cheapestRate
              ? {
                  currency: cheapestRate.currency,
                  base_price: parseFloat(cheapestRate.price_base),
                  taxes: parseFloat(cheapestRate.price_taxes),
                  total_price: parseFloat(cheapestRate.price_total),
                  per_night: parseFloat(cheapestRate.price_per_night),
                  refundable: cheapestRate.refundable,
                  free_cancellation: cheapestRate.free_cancellation,
                }
              : null,
          pricing_available: pricingAvailable,
          room_count: rates.length,
          search_params: {
            check_in,
            check_out,
            nights,
            adults: parseInt(adults),
            children: parseInt(children),
            rooms: parseInt(rooms),
          },
        };
      }),
    );

    res.json({
      success: true,
      hotels: enrichedHotels,
      total: hotelRecords.length,
      pricing_available: !tboError,
      message: tboError ? "Live pricing temporarily unavailable" : "Success",
    });
  } catch (error) {
    console.error("❌ Hotel search error:", error.message);
    res.status(500).json({
      success: false,
      error: "Hotel search failed",
      hotels: [],
    });
  }
});

/**
 * ============================================================================
 * ENDPOINT 3: GET /api/hotels/:propertyId
 * Get detailed information about a specific hotel
 * ============================================================================
 */
router.get("/:propertyId", async (req, res) => {
  try {
    const { propertyId } = req.params;
    const { include_images = true, include_reviews = false } = req.query;

    // Fetch hotel master data
    const hotelResult = await db.query(
      `SELECT * FROM hotel_unified
       WHERE property_id = $1 AND supplier_code = $2`,
      [propertyId, USE_SUPPLIER_FILTER],
    );

    if (hotelResult.rows.length === 0) {
      return res.status(404).json({
        success: false,
        error: "Hotel not found",
      });
    }

    const hotel = hotelResult.rows[0];

    // Fetch images
    let images = [];
    if (include_images === "true" || include_images === true) {
      try {
        const imgResult = await db.query(
          `SELECT image_url, "order" FROM hotel_images
           WHERE property_id = $1
           ORDER BY "order" ASC
           LIMIT 20`,
          [propertyId],
        );
        images = imgResult.rows
          ? imgResult.rows.map((r) => ({
              url: r.image_url,
              order: r.order,
            }))
          : [];
      } catch (error) {
        // Fallback to thumbnail
        if (hotel.thumbnail_url) {
          images = [{ url: hotel.thumbnail_url, order: 0 }];
        }
      }
    }

    // Format response
    res.json({
      success: true,
      hotel: {
        property_id: hotel.property_id,
        supplier_code: hotel.supplier_code,
        supplier_hotel_id: hotel.supplier_hotel_id,
        hotel_name: hotel.hotel_name,
        address: hotel.address,
        city: hotel.city,
        country: hotel.country,
        postal_code: hotel.postal_code,
        star_rating: parseFloat(hotel.star_rating) || null,
        review_score: parseFloat(hotel.review_score) || null,
        review_count: hotel.review_count || 0,
        images: images,
        amenities: hotel.amenities_json || [],
        location: {
          lat: parseFloat(hotel.lat) || null,
          lng: parseFloat(hotel.lng) || null,
          district: hotel.district,
          zone: hotel.zone,
          neighborhood: hotel.neighborhood,
        },
        checkin_from: hotel.checkin_from,
        checkout_until: hotel.checkout_until,
        chain_code: hotel.chain_code,
        brand_code: hotel.brand_code,
        giata_id: hotel.giata_id,
      },
    });
  } catch (error) {
    console.error("❌ Hotel details error:", error.message);
    res.status(500).json({
      success: false,
      error: "Failed to fetch hotel details",
    });
  }
});

/**
 * ============================================================================
 * ENDPOINT 4: POST /api/hotels/:propertyId/rates
 * Get available room rates for a specific hotel
 * ============================================================================
 */
router.post("/:propertyId/rates", async (req, res) => {
  try {
    const { propertyId } = req.params;
    const {
      check_in,
      check_out,
      adults = 2,
      children = 0,
      rooms = 1,
      preferred_currency = "INR",
      refresh = false, // Force refresh from TBO
    } = req.body;

    // Validation
    if (!check_in || !check_out) {
      return res.status(400).json({
        success: false,
        error: "Missing required fields: check_in, check_out",
      });
    }

    // Verify dates
    const checkInDate = new Date(check_in);
    const checkOutDate = new Date(check_out);
    if (isNaN(checkInDate.getTime()) || isNaN(checkOutDate.getTime())) {
      return res.status(400).json({
        success: false,
        error: "Invalid date format. Use YYYY-MM-DD",
      });
    }
    if (checkInDate >= checkOutDate) {
      return res.status(400).json({
        success: false,
        error: "Check-in must be before check-out",
      });
    }

    // Fetch hotel for supplier_hotel_id
    const hotelResult = await db.query(
      `SELECT property_id, supplier_hotel_id, hotel_name, city 
       FROM hotel_unified
       WHERE property_id = $1 AND supplier_code = $2`,
      [propertyId, USE_SUPPLIER_FILTER],
    );

    if (hotelResult.rows.length === 0) {
      return res.status(404).json({
        success: false,
        error: "Hotel not found",
      });
    }

    const hotel = hotelResult.rows[0];

    // Try cache first (unless refresh requested)
    let rates = [];
    let fromCache = false;

    if (!refresh) {
      try {
        const cacheResult = await db.query(
          `SELECT offer_id, property_id, supplier_code, room_name, board_basis, bed_type, 
                  refundable, free_cancellation, occupancy_adults, occupancy_children,
                  currency, price_base, price_taxes, price_total, price_per_night,
                  rate_key_or_token, inclusions_json, cancellable_until
           FROM room_offer_unified
           WHERE property_id = $1 
           AND supplier_code = $2
           AND search_checkin = $3::date
           AND search_checkout = $4::date
           AND expires_at > NOW()
           ORDER BY price_total ASC`,
          [propertyId, USE_SUPPLIER_FILTER, check_in, check_out],
        );

        rates = cacheResult.rows || [];
        fromCache = rates.length > 0;
        console.log(
          `✅ Found ${rates.length} cached rates for hotel ${propertyId}`,
        );
      } catch (error) {
        console.error("⚠️ Cache lookup failed:", error.message);
      }
    }

    // If no cached rates and not explicitly refresh, try TBO
    let tboError = null;
    if (rates.length === 0) {
      try {
        const adapter = getTboAdapter();
        const nights = Math.ceil(
          (checkOutDate - checkInDate) / (1000 * 60 * 60 * 24),
        );

        // Call getHotelRoom or similar enrichment method
        const tboRates = await adapter.getHotelRoom({
          hotelCode: hotel.supplier_hotel_id,
          checkInDate: check_in,
          checkOutDate: check_out,
          nights: nights,
          adults: parseInt(adults),
          children: parseInt(children),
          rooms: parseInt(rooms),
          guestNationality: "IN",
          preferredCurrency: preferred_currency,
        });

        if (Array.isArray(tboRates)) {
          const expiresAt = getExpiryTimestamp();

          // Store in cache
          for (const tboRate of tboRates) {
            try {
              await db.query(
                `INSERT INTO room_offer_unified 
                 (offer_id, property_id, supplier_code, room_name, board_basis, bed_type,
                  refundable, free_cancellation, occupancy_adults, occupancy_children,
                  currency, price_base, price_taxes, price_total, price_per_night,
                  rate_key_or_token, search_checkin, search_checkout, hotel_name, city, 
                  inclusions_json, cancellable_until, expires_at, created_at)
                 VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15,
                         $16, $17::date, $18::date, $19, $20, $21, $22, $23, NOW())
                 ON CONFLICT (offer_id) DO UPDATE SET expires_at = $23`,
                [
                  uuidv4(),
                  propertyId,
                  USE_SUPPLIER_FILTER,
                  tboRate.roomName || tboRate.RoomTypeName,
                  tboRate.boardBasis || tboRate.BoardType || "RO",
                  tboRate.bedType || "Double",
                  tboRate.refundable !== undefined ? tboRate.refundable : true,
                  tboRate.freeCancellation || false,
                  parseInt(adults),
                  parseInt(children),
                  preferred_currency,
                  parseFloat(tboRate.priceBase || tboRate.baseFare || 0),
                  parseFloat(tboRate.priceTaxes || tboRate.taxTotal || 0),
                  parseFloat(tboRate.priceTotal || tboRate.totalFare || 0),
                  parseFloat(tboRate.pricePerNight || tboRate.perNight || 0),
                  tboRate.rateKey || tboRate.RateKey || null,
                  check_in,
                  check_out,
                  hotel.hotel_name,
                  hotel.city,
                  tboRate.inclusions || tboRate.Inclusions || null,
                  tboRate.cancellableUntil || null,
                  expiresAt,
                ],
              );
            } catch (insertError) {
              console.warn("⚠️ Failed to cache rate:", insertError.message);
            }
          }

          rates = tboRates.map((tr) => ({
            offer_id: uuidv4(),
            room_name: tr.roomName || tr.RoomTypeName,
            board_basis: tr.boardBasis || tr.BoardType || "RO",
            bed_type: tr.bedType || "Double",
            refundable: tr.refundable,
            free_cancellation: tr.freeCancellation || false,
            occupancy_adults: parseInt(adults),
            occupancy_children: parseInt(children),
            currency: preferred_currency,
            price_base: parseFloat(tr.priceBase || 0),
            price_taxes: parseFloat(tr.priceTaxes || 0),
            price_total: parseFloat(tr.priceTotal || 0),
            price_per_night: parseFloat(tr.pricePerNight || 0),
            rate_key_or_token: tr.rateKey,
            inclusions_json: tr.inclusions,
            cancellable_until: tr.cancellableUntil,
          }));
        }
      } catch (error) {
        tboError = error;
        console.error("⚠️ TBO rate fetch failed:", error.message);
      }
    }

    // Format response
    const responseRates = rates.map((rate) => ({
      offer_id: rate.offer_id,
      room_name: rate.room_name,
      board_basis: rate.board_basis,
      bed_type: rate.bed_type,
      refundable: rate.refundable,
      free_cancellation: rate.free_cancellation,
      occupancy: {
        adults: rate.occupancy_adults,
        children: rate.occupancy_children,
      },
      pricing: {
        currency: rate.currency,
        base: parseFloat(rate.price_base),
        taxes: parseFloat(rate.price_taxes),
        total: parseFloat(rate.price_total),
        per_night: parseFloat(rate.price_per_night),
      },
      rate_key: rate.rate_key_or_token,
      inclusions: rate.inclusions_json,
      cancellable_until: rate.cancellable_until,
    }));

    res.json({
      success: true,
      property_id: propertyId,
      hotel_name: hotel.hotel_name,
      check_in,
      check_out,
      rates: responseRates,
      total_rooms: responseRates.length,
      pricing_available: responseRates.length > 0,
      from_cache: fromCache,
      message:
        tboError && rates.length === 0
          ? "Pricing temporarily unavailable"
          : "Success",
    });
  } catch (error) {
    console.error("❌ Hotel rates error:", error.message);
    res.status(500).json({
      success: false,
      error: "Failed to fetch room rates",
    });
  }
});

module.exports = router;